requirements.txt
langchain
langchain-openai
pydantic
python-dotenv
langchain-google-genai
langchain-groq



